Re-added Hogstom listing
Please put it in the top left spot on the first sales page (take out the $1,800,000 Glaze house)

Please put a NEW PRICE banner on Hogstrom (it was $425,000) and put $385,000 for the new price at the bottom of the copy.   I will have to do a new PDF for you to attach, and I have to try and remember how to do it.