Pricing changes
Page 1
==============
bottom left Edgemont condo,  change the price from $150,000 to $145,000

bottom right Contemporary - change the price to $305,000  (no banner or anything).

Page 2
==============

Top left - Put a Reduced banner on it and change the price to $225,000

Take out the top right - Pico Condo and insert this one:

ROBINWOOD CONTEMPORARY. REDUCED TO ................................$325,000

Take out the bottom left - Woods Cluster Home.

and move the Hemlock Ridge over to the left - so there are only three on this page.

Here is the last one to fill in page 2.  I forgot to send you a photo of the emergency listing.

WOW !   8 BEDROOMS, 3 BATHS